requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
pycryptodome==3.14.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.6.0
